Refactor: Rename package name from space.taran to dev.arkbuilders (#28)
