+++ app/src/main/java/dev/arkbuilders/arkmemo/data/ResourceMeta.kt
@@ -1,0 +1,12 @@
+package dev.arkbuilders.arkmemo.data
+
+import dev.arkbuilders.arklib.ResourceId
+import java.nio.file.attribute.FileTime
+
+data class ResourceMeta(
+    val id: ResourceId,
+    val name: String,
+    val extension: String,
+    val modified: FileTime,
+    val size: Long
+    )
