+++ app/src/main/java/dev/arkbuilders/arkmemo/preferences/MemoPreferences.kt
@@ -1,0 +1,11 @@
+package dev.arkbuilders.arkmemo.preferences
+
+import java.nio.file.Path
+
+interface MemoPreferences {
+    fun storePath(path: String)
+
+    fun getPathString(): String?
+
+    fun getPath(): Path?
+}
