+++ app/src/main/java/dev/arkbuilders/arkmemo/ui/fragments/NotesFragment.kt
@@ -1,0 +1,100 @@
+package dev.arkbuilders.arkmemo.ui.fragments
+
+import android.os.Bundle
+import android.view.View
+import android.widget.Button
+import android.widget.Toast
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import by.kirich1409.viewbindingdelegate.viewBinding
+import com.google.android.material.floatingactionbutton.FloatingActionButton
+import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.launch
+import dev.arkbuilders.arkmemo.R
+import dev.arkbuilders.arkmemo.ui.viewmodels.NotesViewModel
+import dev.arkbuilders.arkmemo.databinding.FragmentTextNotesBinding
+import dev.arkbuilders.arkmemo.models.TextNote
+import dev.arkbuilders.arkmemo.ui.activities.MainActivity
+import dev.arkbuilders.arkmemo.ui.activities.getTextFromClipBoard
+import dev.arkbuilders.arkmemo.ui.activities.replaceFragment
+import dev.arkbuilders.arkmemo.ui.adapters.TextNotesListAdapter
+
+@AndroidEntryPoint
+class TextNotesFragment: Fragment(R.layout.fragment_text_notes) {
+
+    private val binding by viewBinding(FragmentTextNotesBinding::bind)
+
+    private val activity: MainActivity by lazy {
+        requireActivity() as MainActivity
+    }
+
+    private val textNotesViewModel: NotesViewModel by activityViewModels()
+
+    private lateinit var newNoteButton: FloatingActionButton
+    private lateinit var pasteNoteButton: Button
+
+    private lateinit var recyclerView: RecyclerView
+
+    private val newNoteClickListener = View.OnClickListener {
+        activity.fragment = EditTextNotesFragment()
+        activity.replaceFragment(activity.fragment, EditTextNotesFragment.TAG)
+    }
+
+    private val pasteNoteClickListener = View.OnClickListener {
+        val clipBoardText = requireContext().getTextFromClipBoard()
+        if (clipBoardText != null) {
+            activity.fragment = EditTextNotesFragment.newInstance(clipBoardText)
+            activity.replaceFragment(activity.fragment, EditTextNotesFragment.TAG)
+        }
+        else Toast.makeText(requireContext(), getString(R.string.nothing_to_paste), Toast.LENGTH_SHORT).show()
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        textNotesViewModel.init()
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        recyclerView = binding.include.recyclerView
+        activity.title = getString(R.string.app_name)
+        activity.supportActionBar?.setDisplayHomeAsUpEnabled(false)
+        newNoteButton = binding.newNote
+        pasteNoteButton = binding.pasteNote
+        newNoteButton.setOnClickListener(newNoteClickListener)
+        pasteNoteButton.setOnClickListener(pasteNoteClickListener)
+        lifecycleScope.launch {
+            viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                textNotesViewModel.getTextNotes {
+                    val adapter = TextNotesListAdapter(it)
+                    val layoutManager = LinearLayoutManager(requireContext())
+                    adapter.setActivity(activity)
+                    adapter.setFragmentManager(childFragmentManager)
+                    recyclerView.apply {
+                        this.layoutManager = layoutManager
+                        this.adapter = adapter
+                    }
+                }
+            }
+        }
+    }
+
+    override fun onResume() {
+        super.onResume()
+        activity.fragment = this
+    }
+
+    companion object {
+        const val TAG = "Text Notes Fragment"
+    }
+}
+
+fun Fragment.deleteNote(note: TextNote){
+    val viewModel: NotesViewModel by activityViewModels()
+    viewModel.onDelete(note)
+}
