+++ app/src/main/java/dev/arkbuilders/arkmemo/utils/Constants.kt
@@ -1,0 +1,5 @@
+package dev.arkbuilders.arkmemo.utils
+
+// shared preferences
+const val CRASH_REPORT_ENABLE = "crash_report_enable"
+
