--- app/src/main/java/space/taran/arkmemo/di/PreferencesModule.kt
@@ -1,15 +1,0 @@
-package space.taran.arkmemo.di
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import space.taran.arkmemo.preferences.MemoPreferences
-import space.taran.arkmemo.preferences.MemoPreferencesImpl
-
-@InstallIn(SingletonComponent::class)
-@Module
-abstract class PreferencesModule {
-    @Binds
-    abstract fun bindMemoPreferences(impl: MemoPreferencesImpl): MemoPreferences
-}
