--- app/src/main/java/space/taran/arkmemo/di/RepositoryModule.kt
@@ -1,16 +1,0 @@
-package space.taran.arkmemo.di
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import space.taran.arkmemo.data.repositories.TextNotesRepo
-import space.taran.arkmemo.data.repositories.TextNotesRepoImpl
-
-
-@InstallIn(SingletonComponent::class)
-@Module
-abstract class RepositoryModule {
-    @Binds
-    abstract fun bindRepository(impl: TextNotesRepoImpl): TextNotesRepo
-}
