--- app/src/main/java/space/taran/arkmemo/models/TextNote.kt
@@ -1,20 +1,0 @@
-package space.taran.arkmemo.models
-
-import android.os.Parcelable
-import kotlinx.parcelize.IgnoredOnParcel
-import space.taran.arkmemo.data.ResourceMeta
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class TextNote (
-    val content: Content,
-    @IgnoredOnParcel
-    var meta: ResourceMeta? = null
-): Parcelable
-{
-    @Parcelize
-    data class Content(
-        val title: String,
-        val data: String
-        ): Parcelable
-}
