--- app/src/main/java/space/taran/arkmemo/preferences/MemoPreferences.kt
@@ -1,11 +1,0 @@
-package space.taran.arkmemo.preferences
-
-import java.nio.file.Path
-
-interface MemoPreferences {
-    fun storePath(path: String)
-
-    fun getPathString(): String?
-
-    fun getPath(): Path?
-}
