--- app/src/main/java/space/taran/arkmemo/utils/Constants.kt
@@ -1,5 +1,0 @@
-package space.taran.arkmemo.space.taran.arkmemo.utils
-
-// shared preferences
-const val CRASH_REPORT_ENABLE = "crash_report_enable"
-
